refactor: reorganize and clean
